Feature/force extend (#44)
* 20240418

* 20240418b

* Fix deploy issue - library dependencies

* Fix iBorrowPositionToken for timestamps as uint40

* 20240419

* 20240419b

* 20240420

* 20240422

* 20240422b

* ensure ltv is uint16

* update ltv and apy to uint16

* update ltv and apy to uint16 - 2

* 20240502

* ForceExtendBorrow working

+++ packages/hardhat/contracts/libraries/types/DataTypes.sol
@@ -42,20 +42,20 @@ library DataTypes {
 
     struct BorrowData {
         uint40 startDate;  // Max Date with uint40 is 2106 (83 years from now)
-        uint40 endDate;
-        uint256 principal;
-        uint256 collateral;
-        uint32 apy;
+        uint40 endDate;  // End Date of the loan (uint40)
+        uint256 principal;  // USDC Amount borrowed (6 decimals)
+        uint256 collateral;  // ETH provided as collateral (Wad)
+        uint16 apy;  // Interest rate of loan (percentage)
     }
 
     struct BorrowInternalParams {
         uint256 principal; // Amount of USDC with 6 decimal places
         uint256 collateral; // Amount of ETH collateral with 18 decimal places
-        uint32 ltv; // ltv expressed as a percentage
-        uint40 timestamp;
-        uint40 startDate;
-        address beneficiary;
-        address borrower;
+        uint16 ltv; // ltv expressed as a percentage
+        uint40 timestamp;  // End date of the borrow
+        uint40 startDate;  // Start date of the borrow
+        address beneficiary;  // Account to receive the USDC borrowed
+        address borrower;  // Account to take passession of the BPT
     }
 
     struct calcLPIncreasesParams {
